adding CRUD using DTO

+++ hr-service/src/main/java/tn/esprit/Repository/ProfileRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.Repository;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.Entity.Profile;
 
-public interface ProfileRepository extends CrudRepository<Profile, Integer> {
+public interface ProfileRepository extends CrudRepository<Profile, Long> {
 }
